WIP
This fixes when 64bit values get used as 32bit values.

When possible I've changed our type to the 64bit value.

In cases where our API returns the wrong type I assert if it is too large.

In the majority of cases we simply have to use a int as it is public API
of glib. In all other cases the value gets clamped to the min or max int
value. All of these cases seem extremely unlikely in the first place.

+++ libsoup/auth/soup-auth-manager.c
@@ -272,7 +272,7 @@ soup_auth_manager_extract_challenges (const char *challenges, const char *scheme
 {
         GPtrArray *challenge_list = g_ptr_array_new ();
 	GSList *items, *i, *next;
-	int schemelen = strlen (scheme);
+	gsize schemelen = strlen (scheme);
 	char *item;
 	GString *challenge;
